issue #2400 WIP
